requirements.txt
pygame
python-dotenv
google-genai
python-chess